logout possible nullpointer fixed

+++ app/Http/Controllers/AuthController.php
@@ -66,7 +66,10 @@ class AuthController extends Controller
     }
 
     public function logout(Request $request){
-        $request->user()->tokens()->delete();
+        $user = $request->user();
+        if ( $user && !$user->tokens()->isEmpty()) {
+            $request->user()->tokens()->delete();
+        }
         Auth::guard('web')->logout();
 /*
         Auth::logout();
